v1

+++ DocumentAnalyzerAPI/DTOs/APInvoiceLineDTO.cs
@@ -6,7 +6,8 @@ namespace DocumentAnalyzerAPI.DTOs
     {
         public string? Description { get; set; }
         public decimal? Tax { get; set; }
-        public int?  Quantity { get; set; }
+        public Double?  Quantity { get; set; }
+
         public decimal?  UnitPrice { get; set; }
         public Guid?  ExpenseCategoryId { get; set; }
         public decimal?  DepreciationRatePerYear { get; set; }
